requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
httpx==0.27.2
requests>=2.31.0
qrcode[pil]==7.4.2
Pillow==10.4.0
sqlalchemy==2.0.36
asyncpg==0.29.0
aiosqlite==0.20.0
python-jose==3.3.0
passlib[bcrypt]==1.7.4


